+++ src/literalizer/languages/nim.py
@@ -535,6 +535,7 @@ class Nim(metaclass=LanguageCls):
     pygments_name = "nim"
     supports_special_floats = True
     supports_variable_names = True
+    supports_no_variable_wrap_in_file = False
     dict_supports_heterogeneous_values = False
     supports_dotted_calls = True
     has_free_function_calls = True