+++ src/literalizer/languages/nix.py
@@ -185,6 +185,7 @@ class Nix(metaclass=LanguageCls):
     pygments_name = "nix"
     supports_special_floats = True
     supports_variable_names = True
+    supports_no_variable_wrap_in_file = True
     dict_supports_heterogeneous_values = True
     supports_dotted_calls = True
     has_free_function_calls = True