+++ src/literalizer/languages/objective_c.py
@@ -275,6 +275,7 @@ class ObjectiveC(metaclass=LanguageCls):
     pygments_name = "objective-c"
     supports_special_floats = True
     supports_variable_names = True
+    supports_no_variable_wrap_in_file = False
     dict_supports_heterogeneous_values = True
     supports_dotted_calls = True
     has_free_function_calls = True