+++ src/literalizer/languages/powershell.py
@@ -164,6 +164,7 @@ class PowerShell(metaclass=LanguageCls):
     pygments_name = "powershell"
     supports_special_floats = True
     supports_variable_names = True
+    supports_no_variable_wrap_in_file = True
     dict_supports_heterogeneous_values = True
     supports_dotted_calls = True
     has_free_function_calls = True