+++ src/literalizer/languages/python.py
@@ -542,6 +542,7 @@ class Python(metaclass=LanguageCls):
     pygments_name = "python"
     supports_special_floats = True
     supports_variable_names = True
+    supports_no_variable_wrap_in_file = True
     dict_supports_heterogeneous_values = True
     supports_dotted_calls = True
     has_free_function_calls = True