+++ src/literalizer/languages/roc.py
@@ -509,6 +509,7 @@ class Roc(metaclass=LanguageCls):
     extension = ".roc"
     pygments_name = "text"
     supports_variable_names = True
+    supports_no_variable_wrap_in_file = False
     dict_supports_heterogeneous_values = True
     supports_dotted_calls = True
     has_free_function_calls = True